+++ demo/mario.js
@@ -1,9 +1,4 @@
-kaboom({
-	width: 640,
-	height: 480,
-	stretch: true,
-	letterbox: true,
-});
+kaboom();
 
 // load assets
 loadSprite("bean", "sprites/bean.png");